add microbenchmark for exchange and window function (#5137)
close pingcap/tiflash#4276, close pingcap/tiflash#5138
