+++ dbms/src/Flash/Coprocessor/DAGExpressionAnalyzer.h
@@ -153,7 +153,9 @@ public:
         const tipb::Window & window,
         size_t window_columns_start_index);
 
+#ifndef DBMS_PUBLIC_GTEST
 private:
+#endif
     NamesAndTypes buildOrderColumns(
         const ExpressionActionsPtr & actions,
         const ::google::protobuf::RepeatedPtrField<tipb::ByItem> & order_by);
